+++ functions.php
@@ -7,19 +7,17 @@
  * Version: 4.0.0
  * Text Domain: eightshift-boilerplate
  *
- * @package Eightshift_Boilerplate
- *
- * @since 1.0.0
+ * @package EightshiftBoilerplate
  */
 
 declare( strict_types=1 );
 
-namespace Eightshift_Boilerplate;
+namespace EightshiftBoilerplate;
+
+use EightshiftBoilerplate\Core\Main;
 
 /**
  * If this file is called directly, abort.
- *
- * @since 1.0.0
  */
 if ( ! \defined( 'WPINC' ) ) {
   die;
@@ -27,8 +25,6 @@ if ( ! \defined( 'WPINC' ) ) {
 
 /**
  * Include the autoloader so we can dynamically include the rest of the classes.
- *
- * @since 1.0.0
  */
 require __DIR__ . '/vendor/autoload.php';
 
@@ -38,7 +34,5 @@ require __DIR__ . '/vendor/autoload.php';
  * Since everything within the theme is registered via hooks,
  * then kicking off the theme from this point in the file does
  * not affect the page life cycle.
- *
- * @since 1.0.0
  */
-( new Core\Main() )->register();
+( new Main() )->register();
