+++ page.php
@@ -2,9 +2,7 @@
 /**
  * Display regular page
  *
- * @package Eightshift_Boilerplate
- *
- * @since 1.0.0
+ * @package EightshiftBoilerplate
  */
 
 get_header();
