+++ src/Config.php
@@ -5,28 +5,24 @@
  * A class definition that includes attributes and functions used across both the
  * public side of the site and the admin area.
  *
- * @package Eightshift_Boilerplate\Core
+ * @package EightshiftBoilerplate\Core
  */
 
 declare( strict_types=1 );
 
-namespace Eightshift_Boilerplate\Core;
+namespace EightshiftBoilerplate\Core;
 
-use Eightshift_Libs\Core\Config as Lib_Config;
+use EightshiftBoilerplateVendor\EightshiftLibs\Core\AbstractConfig;
 
 /**
  * The project config class.
- *
- * @since 4.0.0
  */
-class Config extends Lib_Config {
+class Config extends AbstractConfig {
 
   /**
    * Method that returns project name.
    *
    * Generally used for naming assets handlers, languages, etc.
-   *
-   * @since 4.0.0 Added in the project
    */
   public static function get_project_name() : string {
     return 'eightshift-boilerplate';
@@ -36,8 +32,6 @@ class Config extends Lib_Config {
    * Method that returns project version.
    *
    * Generally used for versioning asset handlers while enqueueing them.
-   *
-   * @since 4.0.0 Added in the project
    */
   public static function get_project_version() : string {
     return '1.0.0';
@@ -49,8 +43,6 @@ class Config extends Lib_Config {
    * The WordPress filters live in a global namespace, so we need to prefix them to avoid naming collisions.
    *
    * @return string Full path to asset.
-   *
-   * @since 4.0.0 Added in the project
    */
   public static function get_project_prefix() : string {
     return 'eb';
@@ -66,8 +58,6 @@ class Config extends Lib_Config {
    * defined in wp-config.php
    *
    * @return string Current project environment string.
-   *
-   * @since 4.0.0 Added in the project
    */
   public static function get_project_env() : string {
     if ( defined( 'EB_ENV' ) ) {
@@ -81,8 +71,6 @@ class Config extends Lib_Config {
    * Method that returns project REST-API namespace.
    *
    * Used for namespacing projects REST-API routes and fields.
-   *
-   * @since 4.0.0 Added in the project
    */
   public static function get_project_routes_namespace() : string {
     return static::get_project_name();
@@ -92,8 +80,6 @@ class Config extends Lib_Config {
    * Method that returns project REST-API version.
    *
    * Used for versioning projects REST-API routes and fields.
-   *
-   * @since 4.0.0 Added in the project
    */
   public static function get_project_routes_version() : string {
     return 'v1';
@@ -103,8 +89,6 @@ class Config extends Lib_Config {
    * Method that returns project primary color.
    *
    * Used for styling the mobile browser color and splash screens. Check head.php for more details.
-   *
-   * @since 4.0.0 Added in the project
    */
   public static function get_project_primary_color() : string {
     return '#900000';
@@ -118,8 +102,6 @@ class Config extends Lib_Config {
    * @param string $path Additional path to add to project path.
    *
    * @return string
-   *
-   * @since 4.0.0 Added in the project
    */
   public static function get_project_path( string $path = '' ) : string {
     return get_template_directory() . $path;
