+++ src/admin/FinalModifyAdminAppearance.php
@@ -2,23 +2,21 @@
 /**
  * Modify WordPress admin behavior
  *
- * @package Eightshift_Boilerplate\Admin
+ * @package EightshiftBoilerplate\Admin
  */
 
 declare( strict_types=1 );
 
-namespace Eightshift_Boilerplate\Admin;
+namespace EightshiftBoilerplate\Admin;
 
-use Eightshift_Libs\Core\Service;
+use EightshiftBoilerplateVendor\EightshiftLibs\Core\ServiceInterface;
 
 /**
  * Class that modifies some administrator appearance
  *
  * Example: Change color based on environment, remove dashboard widgets etc.
- *
- * @since 4.0.0
  */
-final class Modify_Admin_Appearance implements Service {
+final class FinalModifyAdminAppearance implements ServiceInterface {
 
   /**
    * List of admin color schemes.
@@ -35,8 +33,6 @@ final class Modify_Admin_Appearance implements Service {
    * Register all the hooks
    *
    * @return void
-   *
-   * @since 4.0.0
    */
   public function register() {
     \add_filter( 'get_user_option_admin_color', [ $this, 'set_admin_color' ], 10, 0 );
@@ -45,9 +41,7 @@ final class Modify_Admin_Appearance implements Service {
   /**
    * Method that changes admin colors based on environment variable
    *
-   * @return string Modified color scheme.
-   *
-   * @since 4.0.0.
+   * @return string Modified color scheme..
    */
   public function set_admin_color() : string {
     if ( ! \defined( 'EB_ENV' ) || ! isset( self::COLOR_SCHEMES[ EB_ENV ] ) ) {
